+++ src/Connection.java
@@ -1,9 +1,9 @@
 public class Connection {
     City city1;
     City city2;
-    int distance;
+    double distance;
 
-    public Connection(City city1, City city2, int distance) {
+    public Connection(City city1, City city2, double distance) {
         this.city1 = city1;
         this.city2 = city2;
         this.distance = distance;
